feat(hoop-ttb.19): complete extensibility implementation with scripts library and notes MCP integration
Complete the four extensibility systems from plan section §22:

**Scripts Library (api_scripts.rs):**
- Add ScriptLibrary and ScriptStore for in-memory script management
- Add ensure_scripts_dir() to create and seed example scripts
- Add seed_example_script() with hello-world example
- Add start_watcher() for hot-reload on directory changes
- Full integration with daemon's script_library state

**Notes MCP Integration (hoop-mcp/):**
- Add notes module to MCP server (lib.rs, main.rs)
- Add read_note tool to agent's tool belt (tools.rs)
- Agent can now access team conventions, glossaries, and reference material

**Minor fixes:**
- Fix AxumPath -> Path extraction in api_skills.rs for consistency
- Add Debug derive to DetectionState in api_reflection_detection.rs
- Fix utoipa::ToSchema derive attributes in api_onboarding.rs
- Add type annotations for PathBuf in capacity.rs

Scripts are now:
1. Auto-discovered from ~/.hoop/scripts/<name> with optional manifest.yml
2. Executable via API (manual), event subscriptions, or cron schedule
3. Hot-reloaded when directory changes
4. Fully audited with actor identity and args hash

Notes are now:
1. Accessible to agent via read_note() MCP tool
2. Auto-discovered from ~/.hoop/notes/ and project .hoop/notes/
3. Hot-reloaded on file changes
4. Supports YAML frontmatter for metadata

+++ hoop-daemon/src/cost_anomaly.rs
@@ -325,7 +325,7 @@ pub fn check_on_stitch_close(
                 crate::ws::ClosestPatternMatch {
                     pattern_id: p.pattern_id.clone(),
                     pattern_name: p.pattern_name.clone(),
-                    similarity: p.similarity,
+                    similarity: p.similarity as f64,
                     recommended_fix_template_md: p.recommended_fix_template_md.clone(),
                 }
             });
